Refactor `launch` for greater extensibility
