+++ src/accelerate/utils/__init__.py
@@ -99,7 +99,16 @@ if is_deepspeed_available():
         HfDeepSpeedConfig,
     )
 
-from .launch import PrepareForLaunch, _filter_args, get_launch_prefix, prepare_tpu
+from .launch import (
+    PrepareForLaunch,
+    _filter_args,
+    get_launch_prefix,
+    prepare_deepspeed_cmd_env,
+    prepare_multi_gpu_env,
+    prepare_sagemager_args_inputs,
+    prepare_simple_launcher_cmd_env,
+    prepare_tpu,
+)
 from .megatron_lm import (
     AbstractTrainStep,
     BertTrainStep,
